Integration tests: Combine bitcoin wallet tests
